chore: move result table into separate folder
